Revert "Move persistence ontology structures to provenance module"
This reverts commit 4c3a03881ad7f6f33769afb21996096afc28e5cd.
