+++ packages/graph/hash_graph/lib/graph/src/api/rest/entity_type.rs
@@ -18,12 +18,10 @@ use crate::{
     api::rest::{api_resource::RoutedResource, read_from_store, report_to_status_code},
     ontology::{
         domain_validator::{DomainValidator, ValidateOntologyType},
-        patch_id_and_parse,
+        patch_id_and_parse, PersistedEntityType, PersistedOntologyIdentifier,
+        PersistedOntologyMetadata,
     },
-    provenance::{
-        CreatedById, OwnedById, PersistedEntityType, PersistedOntologyIdentifier,
-        PersistedOntologyMetadata, UpdatedById,
-    },
+    provenance::{CreatedById, OwnedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         error::{BaseUriAlreadyExists, BaseUriDoesNotExist},
