+++ packages/graph/hash_graph/lib/graph/src/api/rest/property_type.rs
@@ -19,12 +19,10 @@ use crate::{
     api::rest::{read_from_store, report_to_status_code},
     ontology::{
         domain_validator::{DomainValidator, ValidateOntologyType},
-        patch_id_and_parse,
+        patch_id_and_parse, PersistedOntologyIdentifier, PersistedOntologyMetadata,
+        PersistedPropertyType,
     },
-    provenance::{
-        CreatedById, OwnedById, PersistedOntologyIdentifier, PersistedOntologyMetadata,
-        PersistedPropertyType, UpdatedById,
-    },
+    provenance::{CreatedById, OwnedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         query::Expression, BaseUriAlreadyExists, BaseUriDoesNotExist, PropertyTypeStore, StorePool,
