+++ packages/graph/hash_graph/lib/graph/src/provenance/mod.rs
@@ -1,10 +1,3 @@
 mod ids;
-mod ontology;
 
-pub use self::{
-    ids::{AccountId, CreatedById, OwnedById, RemovedById, UpdatedById},
-    ontology::{
-        PersistedDataType, PersistedEntityType, PersistedLinkType, PersistedOntologyIdentifier,
-        PersistedOntologyMetadata, PersistedPropertyType,
-    },
-};
+pub use self::ids::{AccountId, CreatedById, OwnedById, RemovedById, UpdatedById};
