+++ packages/graph/hash_graph/lib/graph/src/store/postgres/mod.rs
@@ -35,12 +35,11 @@ use crate::{
         Entity, EntityId, KnowledgeGraphQueryDepth, Link, PersistedEntity,
         PersistedEntityIdentifier, PersistedEntityMetadata, PersistedLink,
     },
-    ontology::OntologyQueryDepth,
-    provenance::{
-        AccountId, CreatedById, OwnedById, PersistedDataType, PersistedEntityType,
-        PersistedLinkType, PersistedOntologyIdentifier, PersistedOntologyMetadata,
-        PersistedPropertyType, RemovedById, UpdatedById,
+    ontology::{
+        OntologyQueryDepth, PersistedDataType, PersistedEntityType, PersistedLinkType,
+        PersistedOntologyIdentifier, PersistedOntologyMetadata, PersistedPropertyType,
     },
+    provenance::{AccountId, CreatedById, OwnedById, RemovedById, UpdatedById},
     shared::identifier::{GraphElementIdentifier, LinkId},
     store::{
         error::VersionedUriAlreadyExists,
