+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/link_type/mod.rs
@@ -7,9 +7,8 @@ use tokio_postgres::GenericClient;
 use type_system::{uri::VersionedUri, LinkType};
 
 use crate::{
-    provenance::{
-        CreatedById, OwnedById, PersistedLinkType, PersistedOntologyMetadata, UpdatedById,
-    },
+    ontology::{PersistedLinkType, PersistedOntologyMetadata},
+    provenance::{CreatedById, OwnedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
