+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/read.rs
@@ -4,7 +4,7 @@ use futures::{stream, StreamExt, TryStreamExt};
 use type_system::{DataType, EntityType, LinkType, PropertyType};
 
 use crate::{
-    provenance::{
+    ontology::{
         PersistedDataType, PersistedEntityType, PersistedLinkType, PersistedOntologyIdentifier,
         PersistedOntologyMetadata, PersistedPropertyType,
     },
