+++ packages/graph/hash_graph/tests/integration/postgres/mod.rs
@@ -10,11 +10,11 @@ use std::str::FromStr;
 use error_stack::{Report, Result};
 use graph::{
     knowledge::{Entity, EntityId, Link, PersistedEntity, PersistedEntityMetadata, PersistedLink},
-    provenance::{
-        AccountId, CreatedById, OwnedById, PersistedDataType, PersistedEntityType,
-        PersistedLinkType, PersistedOntologyMetadata, PersistedPropertyType, RemovedById,
-        UpdatedById,
+    ontology::{
+        PersistedDataType, PersistedEntityType, PersistedLinkType, PersistedOntologyMetadata,
+        PersistedPropertyType,
     },
+    provenance::{AccountId, CreatedById, OwnedById, RemovedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         error::LinkRemovalError,
